edit, update, delete items

+++ db/migrate/20151227003242_create_items.rb
@@ -6,8 +6,8 @@ class CreateItems < ActiveRecord::Migration
       t.string :kit, null: false
       t.string :item_num, null: false
       t.string :item_desc, null: false
-      t.decimal :item_qty, null: false
-      t.decimal :item_price, null: false
+      t.integer :item_qty, null: false
+      t.float :item_price, null: false
       t.string :item_price_type, null: false
       t.timestamps null: false
     end
